8260467: Move well-known classes from systemDictionary.hpp to vmClasses.hpp
