+++ src/hotspot/cpu/s390/methodHandles_s390.cpp
@@ -27,6 +27,7 @@
 #include "jvm.h"
 #include "asm/macroAssembler.inline.hpp"
 #include "classfile/javaClasses.inline.hpp"
+#include "classfile/vmClasses.hpp"
 #include "interpreter/interpreter.hpp"
 #include "logging/log.hpp"
 #include "logging/logStream.hpp"
@@ -56,7 +57,7 @@ static RegisterOrConstant constant(int value) {
 void MethodHandles::load_klass_from_Class(MacroAssembler* _masm, Register klass_reg,
                                           Register temp_reg, Register temp2_reg) {
   if (VerifyMethodHandles) {
-    verify_klass(_masm, klass_reg, SystemDictionary::WK_KLASS_ENUM_NAME(java_lang_Class),
+    verify_klass(_masm, klass_reg, VM_CLASS_ID(java_lang_Class),
                  temp_reg, temp2_reg, "MH argument is a Class");
   }
   __ z_lg(klass_reg, Address(klass_reg, java_lang_Class::klass_offset()));
@@ -75,12 +76,12 @@ static int check_nonzero(const char* xname, int x) {
 
 #ifdef ASSERT
 void MethodHandles::verify_klass(MacroAssembler* _masm,
-                                 Register obj_reg, SystemDictionary::WKID klass_id,
+                                 Register obj_reg, VMClassID klass_id,
                                  Register temp_reg, Register temp2_reg,
                                  const char* error_message) {
 
-  InstanceKlass** klass_addr = SystemDictionary::well_known_klass_addr(klass_id);
-  Klass* klass = SystemDictionary::well_known_klass(klass_id);
+  InstanceKlass** klass_addr = VMClassses::klass_addr_at(klass_id);
+  Klass* klass = VMClassses::klass_at(klass_id);
 
   assert(temp_reg != Z_R0 && // Is used as base register!
          temp_reg != noreg && temp2_reg != noreg, "need valid registers!");
@@ -387,7 +388,7 @@ void MethodHandles::generate_method_handle_dispatch(MacroAssembler* _masm,
   if (VerifyMethodHandles) {
     // Make sure the trailing argument really is a MemberName (caller responsibility).
     verify_klass(_masm, member_reg,
-                 SystemDictionary::WK_KLASS_ENUM_NAME(MemberName_klass),
+                 VM_CLASS_ID(MemberName_klass),
                  temp1, temp2,
                  "MemberName required for invokeVirtual etc.");
   }
