+++ src/hotspot/share/classfile/systemDictionary.cpp
@@ -43,6 +43,7 @@
 #include "classfile/stringTable.hpp"
 #include "classfile/symbolTable.hpp"
 #include "classfile/systemDictionary.hpp"
+#include "classfile/vmClasses.hpp"
 #include "classfile/vmSymbols.hpp"
 #include "code/codeCache.hpp"
 #include "gc/shared/gcTraceTime.inline.hpp"
@@ -57,7 +58,6 @@
 #include "memory/universe.hpp"
 #include "oops/access.inline.hpp"
 #include "oops/instanceKlass.hpp"
-#include "oops/instanceRefKlass.hpp"
 #include "oops/klass.inline.hpp"
 #include "oops/method.inline.hpp"
 #include "oops/objArrayKlass.hpp"
@@ -93,11 +93,6 @@ LoaderConstraintTable* SystemDictionary::_loader_constraints  = NULL;
 ResolutionErrorTable*  SystemDictionary::_resolution_errors   = NULL;
 SymbolPropertyTable*   SystemDictionary::_invoke_method_table = NULL;
 ProtectionDomainCacheTable*   SystemDictionary::_pd_cache_table = NULL;
-
-InstanceKlass*      SystemDictionary::_well_known_klasses[SystemDictionary::WKID_LIMIT]
-                                                          =  { NULL /*, NULL...*/ };
-
-InstanceKlass*      SystemDictionary::_box_klasses[T_VOID+1]      =  { NULL /*, NULL...*/ };
 
 OopHandle   SystemDictionary::_java_system_loader;
 OopHandle   SystemDictionary::_java_platform_loader;
@@ -1418,36 +1413,6 @@ void SystemDictionary::load_shared_class_misc(InstanceKlass* ik, ClassLoaderData
   }
 }
 
-void SystemDictionary::quick_resolve(InstanceKlass* klass, ClassLoaderData* loader_data, Handle domain, TRAPS) {
-  assert(!Universe::is_fully_initialized(), "We can make short cuts only during VM initialization");
-  assert(klass->is_shared(), "Must be shared class");
-  if (klass->class_loader_data() != NULL) {
-    return;
-  }
-
-  // add super and interfaces first
-  Klass* super = klass->super();
-  if (super != NULL && super->class_loader_data() == NULL) {
-    assert(super->is_instance_klass(), "Super should be instance klass");
-    quick_resolve(InstanceKlass::cast(super), loader_data, domain, CHECK);
-  }
-
-  Array<InstanceKlass*>* ifs = klass->local_interfaces();
-  for (int i = 0; i < ifs->length(); i++) {
-    InstanceKlass* ik = ifs->at(i);
-    if (ik->class_loader_data()  == NULL) {
-      quick_resolve(ik, loader_data, domain, CHECK);
-    }
-  }
-
-  klass->restore_unshareable_info(loader_data, domain, NULL, THREAD);
-  load_shared_class_misc(klass, loader_data, CHECK);
-  Dictionary* dictionary = loader_data->dictionary();
-  unsigned int hash = dictionary->compute_hash(klass->name());
-  dictionary->add_klass(hash, klass->name(), klass);
-  add_to_hierarchy(klass);
-  assert(klass->is_loaded(), "Must be in at least loaded state");
-}
 #endif // INCLUDE_CDS
 
 InstanceKlass* SystemDictionary::load_instance_class(Symbol* class_name, Handle class_loader, TRAPS) {
@@ -1857,13 +1822,6 @@ bool SystemDictionary::do_unloading(GCTimer* gc_timer) {
   return unloading_occurred;
 }
 
-// CDS: scan and relocate all classes referenced by _well_known_klasses[].
-void SystemDictionary::well_known_klasses_do(MetaspaceClosure* it) {
-  for (int id = FIRST_WKID; id < WKID_LIMIT; id++) {
-    it->push(well_known_klass_addr((WKID)id));
-  }
-}
-
 void SystemDictionary::methods_do(void f(Method*)) {
   // Walk methods in loaded classes
   MutexLocker ml(ClassLoaderDataGraph_lock);
@@ -1884,189 +1842,11 @@ void SystemDictionary::initialize(TRAPS) {
   _pd_cache_table = new ProtectionDomainCacheTable(defaultProtectionDomainCacheSize);
 
   // Resolve basic classes
-  resolve_well_known_classes(CHECK);
+  vmClasses::resolve_all(CHECK);
   // Resolve classes used by archived heap objects
   if (UseSharedSpaces) {
     HeapShared::resolve_classes(CHECK);
   }
-}
-
-// Compact table of directions on the initialization of klasses:
-// TODO: we should change the base type of vmSymbolID from int to short. Then we can declare this
-// array as vmSymbolID wk_init_info[] anf avoid all the type casts.
-static const short wk_init_info[] = {
-  #define WK_KLASS_INIT_INFO(name, symbol) \
-    ((short)VM_SYMBOL_ENUM_NAME(symbol)),
-
-  WK_KLASSES_DO(WK_KLASS_INIT_INFO)
-  #undef WK_KLASS_INIT_INFO
-  0
-};
-
-#ifdef ASSERT
-bool SystemDictionary::is_well_known_klass(Symbol* class_name) {
-  int sid;
-  for (int i = 0; (sid = wk_init_info[i]) != 0; i++) {
-    Symbol* symbol = vmSymbols::symbol_at(vmSymbols::as_SID(sid));
-    if (class_name == symbol) {
-      return true;
-    }
-  }
-  return false;
-}
-
-bool SystemDictionary::is_well_known_klass(Klass* k) {
-  return is_well_known_klass(k->name());
-}
-#endif
-
-bool SystemDictionary::resolve_wk_klass(WKID id, TRAPS) {
-  assert(id >= (int)FIRST_WKID && id < (int)WKID_LIMIT, "oob");
-  int sid = wk_init_info[id - FIRST_WKID];
-  Symbol* symbol = vmSymbols::symbol_at(vmSymbols::as_SID(sid));
-  InstanceKlass** klassp = &_well_known_klasses[id];
-
-#if INCLUDE_CDS
-  if (UseSharedSpaces && !JvmtiExport::should_post_class_prepare()) {
-    InstanceKlass* k = *klassp;
-    assert(k->is_shared_boot_class(), "must be");
-
-    ClassLoaderData* loader_data = ClassLoaderData::the_null_class_loader_data();
-    quick_resolve(k, loader_data, Handle(), CHECK_false);
-    return true;
-  }
-#endif // INCLUDE_CDS
-
-  if (!is_wk_klass_loaded(*klassp)) {
-    Klass* k = resolve_or_fail(symbol, true, CHECK_false);
-    (*klassp) = InstanceKlass::cast(k);
-  }
-  return ((*klassp) != NULL);
-}
-
-void SystemDictionary::resolve_wk_klasses_until(WKID limit_id, WKID &start_id, TRAPS) {
-  assert((int)start_id <= (int)limit_id, "IDs are out of order!");
-  for (int id = (int)start_id; id < (int)limit_id; id++) {
-    assert(id >= (int)FIRST_WKID && id < (int)WKID_LIMIT, "oob");
-    resolve_wk_klass((WKID)id, CHECK);
-  }
-
-  // move the starting value forward to the limit:
-  start_id = limit_id;
-}
-
-void SystemDictionary::resolve_well_known_classes(TRAPS) {
-  assert(!Object_klass_loaded(), "well-known classes should only be initialized once");
-
-  // Create the ModuleEntry for java.base.  This call needs to be done here,
-  // after vmSymbols::initialize() is called but before any classes are pre-loaded.
-  ClassLoader::classLoader_init2(CHECK);
-
-  // Preload commonly used klasses
-  WKID scan = FIRST_WKID;
-  // first do Object, then String, Class
-#if INCLUDE_CDS
-  if (UseSharedSpaces) {
-    resolve_wk_klasses_through(WK_KLASS_ENUM_NAME(Object_klass), scan, CHECK);
-
-    // It's unsafe to access the archived heap regions before they
-    // are fixed up, so we must do the fixup as early as possible
-    // before the archived java objects are accessed by functions
-    // such as java_lang_Class::restore_archived_mirror and
-    // ConstantPool::restore_unshareable_info (restores the archived
-    // resolved_references array object).
-    //
-    // HeapShared::fixup_mapped_heap_regions() fills the empty
-    // spaces in the archived heap regions and may use
-    // SystemDictionary::Object_klass(), so we can do this only after
-    // Object_klass is resolved. See the above resolve_wk_klasses_through()
-    // call. No mirror objects are accessed/restored in the above call.
-    // Mirrors are restored after java.lang.Class is loaded.
-    HeapShared::fixup_mapped_heap_regions();
-
-    // Initialize the constant pool for the Object_class
-    assert(Object_klass()->is_shared(), "must be");
-    Object_klass()->constants()->restore_unshareable_info(CHECK);
-    resolve_wk_klasses_through(WK_KLASS_ENUM_NAME(Class_klass), scan, CHECK);
-  } else
-#endif
-  {
-    resolve_wk_klasses_through(WK_KLASS_ENUM_NAME(Class_klass), scan, CHECK);
-  }
-
-  assert(WK_KLASS(Object_klass) != NULL, "well-known classes should now be initialized");
-
-  java_lang_Object::register_natives(CHECK);
-
-  // Calculate offsets for String and Class classes since they are loaded and
-  // can be used after this point.
-  java_lang_String::compute_offsets();
-  java_lang_Class::compute_offsets();
-
-  // Fixup mirrors for classes loaded before java.lang.Class.
-  Universe::initialize_basic_type_mirrors(CHECK);
-  Universe::fixup_mirrors(CHECK);
-
-  // do a bunch more:
-  resolve_wk_klasses_through(WK_KLASS_ENUM_NAME(Reference_klass), scan, CHECK);
-
-  // The offsets for jlr.Reference must be computed before
-  // InstanceRefKlass::update_nonstatic_oop_maps is called. That function uses
-  // the offsets to remove the referent and discovered fields from the oop maps,
-  // as they are treated in a special way by the GC. Removing these oops from the
-  // oop maps must be done before the usual subclasses of jlr.Reference are loaded.
-  java_lang_ref_Reference::compute_offsets();
-
-  // Preload ref klasses and set reference types
-  WK_KLASS(Reference_klass)->set_reference_type(REF_OTHER);
-  InstanceRefKlass::update_nonstatic_oop_maps(WK_KLASS(Reference_klass));
-
-  resolve_wk_klasses_through(WK_KLASS_ENUM_NAME(PhantomReference_klass), scan, CHECK);
-  WK_KLASS(SoftReference_klass)->set_reference_type(REF_SOFT);
-  WK_KLASS(WeakReference_klass)->set_reference_type(REF_WEAK);
-  WK_KLASS(FinalReference_klass)->set_reference_type(REF_FINAL);
-  WK_KLASS(PhantomReference_klass)->set_reference_type(REF_PHANTOM);
-
-  // JSR 292 classes
-  WKID jsr292_group_start = WK_KLASS_ENUM_NAME(MethodHandle_klass);
-  WKID jsr292_group_end   = WK_KLASS_ENUM_NAME(VolatileCallSite_klass);
-  resolve_wk_klasses_until(jsr292_group_start, scan, CHECK);
-  resolve_wk_klasses_through(jsr292_group_end, scan, CHECK);
-  WKID last = WKID_LIMIT;
-  resolve_wk_klasses_until(last, scan, CHECK);
-
-  _box_klasses[T_BOOLEAN] = WK_KLASS(Boolean_klass);
-  _box_klasses[T_CHAR]    = WK_KLASS(Character_klass);
-  _box_klasses[T_FLOAT]   = WK_KLASS(Float_klass);
-  _box_klasses[T_DOUBLE]  = WK_KLASS(Double_klass);
-  _box_klasses[T_BYTE]    = WK_KLASS(Byte_klass);
-  _box_klasses[T_SHORT]   = WK_KLASS(Short_klass);
-  _box_klasses[T_INT]     = WK_KLASS(Integer_klass);
-  _box_klasses[T_LONG]    = WK_KLASS(Long_klass);
-  //_box_klasses[T_OBJECT]  = WK_KLASS(object_klass);
-  //_box_klasses[T_ARRAY]   = WK_KLASS(object_klass);
-
-#ifdef ASSERT
-  if (UseSharedSpaces) {
-    JVMTI_ONLY(assert(JvmtiExport::is_early_phase(),
-                      "All well known classes must be resolved in JVMTI early phase"));
-    for (int i = FIRST_WKID; i < last; i++) {
-      InstanceKlass* k = _well_known_klasses[i];
-      assert(k->is_shared(), "must not be replaced by JVMTI class file load hook");
-    }
-  }
-#endif
-}
-
-// Tells if a given klass is a box (wrapper class, such as java.lang.Integer).
-// If so, returns the basic type it holds.  If not, returns T_OBJECT.
-BasicType SystemDictionary::box_klass_type(Klass* k) {
-  assert(k != NULL, "");
-  for (int i = T_BOOLEAN; i < T_VOID+1; i++) {
-    if (_box_klasses[i] == k)
-      return (BasicType)i;
-  }
-  return T_OBJECT;
 }
 
 // Constraints on class loaders. The details of the algorithm can be
@@ -2814,10 +2594,6 @@ ProtectionDomainCacheEntry* SystemDictionary::cache_get(Handle protection_domain
 
 ClassLoaderData* SystemDictionary::class_loader_data(Handle class_loader) {
   return ClassLoaderData::class_loader_data(class_loader());
-}
-
-bool SystemDictionary::is_wk_klass_loaded(InstanceKlass* klass) {
-  return !(klass == NULL || !klass->is_loaded());
 }
 
 bool SystemDictionary::is_nonpublic_Object_method(Method* m) {
