+++ src/hotspot/share/memory/metaspaceShared.cpp
@@ -36,6 +36,7 @@
 #include "classfile/stringTable.hpp"
 #include "classfile/systemDictionary.hpp"
 #include "classfile/systemDictionaryShared.hpp"
+#include "classfile/vmClasses.hpp"
 #include "classfile/vmSymbols.hpp"
 #include "code/codeCache.hpp"
 #include "gc/shared/gcVMOperations.hpp"
@@ -500,7 +501,7 @@ void MetaspaceShared::serialize(SerializeClosure* soc) {
   InstanceMirrorKlass::serialize_offsets(soc);
 
   // Dump/restore well known classes (pointers)
-  SystemDictionaryShared::serialize_well_known_klasses(soc);
+  SystemDictionaryShared::serialize_vm_classes(soc);
   soc->do_tag(--tag);
 
   CppVtables::serialize(soc);
@@ -780,7 +781,7 @@ void VM_PopulateDumpSharedSpace::doit() {
   _open_archive_heap_regions = NULL;
   dump_java_heap_objects();
 
-  builder.relocate_well_known_klasses();
+  builder.relocate_vm_classes();
 
   log_info(cds)("Update method trampolines");
   builder.update_method_trampolines();
@@ -848,7 +849,7 @@ void VM_PopulateDumpSharedSpace::doit() {
   }
 
   // There may be pending VM operations. We have changed some global states
-  // (such as SystemDictionary::_well_known_klasses) that may cause these VM operations
+  // (such as vmClasses::_klasses) that may cause these VM operations
   // to fail. For safety, forget these operations and exit the VM directly.
   vm_direct_exit(0);
 }
